# specs/5-better-auth-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Better Auth Integration API
  description: API specification for Better Auth (JWT) integration with Todo app
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
...
  /api/tasks:
    get:
      summary: Get current user's tasks
      description: Retrieve all tasks belonging to the authenticated user - enforced by database-level row security that filters by authenticated user ID
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved user's tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a new task for current user
      description: Creates a new task assigned to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a task by ID if it belongs to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a specific task
      description: Updates a task if it belongs to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a specific task
      description: Deletes a task if it belongs to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Updates the completion status of a task if it belongs to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  example: true
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth, passed in Authorization header or retrieved from httpOnly cookie
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          example: false
        user_id:
          type: integer
          example: 1
          description: "Owner of the task - enforced by database-level row security"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
    TaskInput:
      type: object
      properties:
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          example: false
      required:
        - title
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Bad Request"
              message:
                type: string
                example: "Invalid input parameters"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication token is missing or invalid"
    Forbidden:
      description: Access to resource denied
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Forbidden"
              message:
                type: string
                example: "You do not have permission to access this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Not Found"
              message:
                type: string
                example: "The requested resource was not found"
...